feat: :sparkles: projects 페이지 scss 및 app 라우터로 마이그레이션 완료
